Refactor des pages Join/Team/
- Ajout de la page Membre (Contient la liste des membres)

- Tout ce qui étaient avant le bouton "rejoins nous" est sur la page Equipe

- La page Rejoins-nous ne contient que la partie FAQ et le bouton "rejoins nous"
