+++ src/data/members.ts
@@ -196,4 +196,8 @@ export const softwareTeam: Member[] = [
     program: 'engineering_software',
     title: 'title_merchandise_manager',
   },
+  {
+    name: 'Ange Olive Dagang Nyappa',
+    program: 'engineering_software',
+  },  
 ];
